[ArrowFlight] Refactor Arrow SDK for better readability/maintability.
