+++ rust/CLAUDE.md
@@ -32,7 +32,7 @@ This is a Cargo workspace. The workspace root is `rust/Cargo.toml`.
 - `headers_provider.rs` — `HeadersProvider` trait + OAuth implementation
 - `landing_zone.rs` — Batches records before sending over gRPC
 - `record_types.rs` — `EncodedRecord`, `ProtoMessage`, `JsonString`
-- `arrow_stream.rs` — Arrow Flight ingestion (Beta; behind `arrow-flight` feature flag)
+- `stream/arrow/` — Arrow Flight ingestion, ACK rotation, connection, and recovery modules (Beta; behind `arrow-flight`)
 
 ## Build commands
