+++ rust/NEXT_CHANGELOG.md
@@ -8,9 +8,19 @@
 
 ### Bug Fixes
 
+- Fixed Arrow Flight recovery sender lifetime: replacement senders are now published
+  only after pending replay succeeds, while initial supervisor handoff and failed or
+  cancelled replay promptly drop redundant senders instead of retaining incomplete
+  `DoPut` request channels until later teardown.
+
 ### Documentation
 
 ### Internal Changes
+
+- Reorganized Arrow Flight under `stream/arrow/` with focused API, connection,
+  ACK, supervisor, and batch modules and no public API changes. Its tracing
+  target now follows the module path:
+  `databricks_zerobus_ingest_sdk::stream::arrow`.
 
 ### Breaking Changes
