+++ rust/sdk/src/builder/stream_builder.rs
@@ -34,7 +34,7 @@ use crate::{
 #[cfg(feature = "arrow-flight")]
 use crate::arrow_configuration::ArrowStreamConfigurationOptions;
 #[cfg(feature = "arrow-flight")]
-use crate::arrow_stream::{ArrowSchema, ArrowTableProperties, ZerobusArrowStream};
+use crate::stream::{ArrowSchema, ArrowTableProperties, ZerobusArrowStream};
 
 /// Internal representation of the authentication configuration.
 enum AuthConfig {
