+++ rust/sdk/src/lib.rs
@@ -41,8 +41,6 @@ pub mod databricks {
 mod arrow_configuration;
 #[cfg(feature = "arrow-flight")]
 mod arrow_metadata;
-#[cfg(feature = "arrow-flight")]
-mod arrow_stream;
 mod builder;
 mod callbacks;
 mod client_warnings;
@@ -66,8 +64,6 @@ mod token_cache;
 
 #[cfg(feature = "arrow-flight")]
 pub use arrow_configuration::ArrowStreamConfigurationOptions;
-#[cfg(feature = "arrow-flight")]
-pub use arrow_stream::{ArrowSchema, DataType, Field, RecordBatch, TimeUnit, ZerobusArrowStream};
 pub use builder::{StreamBuilder, ZerobusSdkBuilder};
 pub use callbacks::AckCallback;
 pub use default_token_factory::DefaultTokenFactory;
@@ -91,6 +87,8 @@ pub use sdk::{ZerobusSdk, DEFAULT_SDK_IDENTIFIER};
 #[cfg(feature = "testing")]
 pub use stream::CallbackHandlerHarness;
 pub use stream::ZerobusStream;
+#[cfg(feature = "arrow-flight")]
+pub use stream::{ArrowSchema, DataType, Field, RecordBatch, TimeUnit, ZerobusArrowStream};
 pub use stream_configuration::StreamConfigurationOptions;
 #[cfg(feature = "testing")]
 pub use tls_config::NoTlsConfig;
