+++ rust/sdk/src/sdk.rs
@@ -16,7 +16,7 @@ use crate::stream::ZerobusStream;
 use crate::{StreamBuilder, TlsConfig, ZerobusError, ZerobusResult, ZerobusSdkBuilder};
 
 #[cfg(feature = "arrow-flight")]
-use crate::arrow_stream::ZerobusArrowStream;
+use crate::stream::ZerobusArrowStream;
 
 /// Default identifier the SDK sends as the HTTP `user-agent` header on every
 /// request. Use [`ZerobusSdkBuilder::application_name`] to append an
